Transfer function management update
Change the -xf flags to it loads a color (1dt or .json) and a possible associated range
This colormap and range impact the loading of subsequent volumes specified
on the command line. This help loading a larger compound volume made of many files.

Also update the .json to support Point entries for opacity control points.